+++ R/arc-select.R
@@ -245,7 +245,7 @@ arc_select <- function(
   # if the result is empty we return a nothing with a message
   if (rlang::is_empty(res)) {
     cli::cli_alert_info("No features returned from query")
-    return(res)
+    return(arcgisutils::fields_as_ptype_df(list_fields(x)))
   }
 
   # we ensure that the CRS is added